// src/catalog/aws/ec2-specs.gen.ts
export interface Ec2Spec {
  readonly arch?: string
  readonly network?: readonly [baseline: number, burst: number]
  readonly ebsMbps?: readonly [baseline: number, max: number]
  readonly ebsIops?: readonly [baseline: number, max: number]
  readonly store?: {
    readonly disks: number
    readonly diskGB: number
    readonly nvme: boolean
    readonly ssd: boolean
    readonly readIops?: number
    readonly writeIops?: number
  }
  readonly maxPods?: number
}

export const ec2SpecsRetrieved = '2026-09-24'

export const ec2Specs: Readonly<Record<string, Ec2Spec>> = {
// ...
  'm7g.2xlarge': { arch: 'arm64', network: [3.75, 15], ebsMbps: [2500, 10000], ebsIops: [12000, 40000], maxPods: 58 },
// ...
  'm7g.xlarge': { arch: 'arm64', network: [1.876, 12.5], ebsMbps: [1250, 10000], ebsIops: [6000, 40000], maxPods: 58 },
// ...
}